requirements.txt
amqp==2.1.1
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
boto3==1.4.1
botocore==1.4.78
bsddb3==6.2.1
celery==4.0.0
certifi==2016.9.26
click==6.6
configparser==3.5.0
cycler==0.10.0
Cython==0.25.1
decorator==4.0.10
docutils==0.12
entrypoints==0.2.2
enum34==1.1.6
eventlet==0.19.0
Flask==0.11.1
functools32==3.2.3.post2
futures==3.0.5
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
ipykernel==4.5.1
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
kombu==4.0.0
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.10
pypinyin==0.15.0
python-dateutil==2.6.0
pytz==2016.7
pyzmq==16.0.2
qtconsole==4.2.1
redis==2.10.5
s3transfer==0.1.9
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
unqlite==0.6.0
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.11
widgetsnbextension==1.2.6
